add reset rot, rm normalize z

+++ src/compute_map_coords.c
@@ -6,13 +6,13 @@
 /*   By: alevra <alevra@student.42lyon.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/17 18:55:23 by alevra            #+#    #+#             */
-/*   Updated: 2023/01/26 03:07:15 by alevra           ###   ########lyon.fr   */
+/*   Updated: 2023/01/30 11:50:21 by alevra           ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "fdf.h"
 
-void	compute_map_coords(t_map *map, int spacing, float height_factor)
+void	compute_map_coords(t_map *map, float spacing, float height_factor)
 {
 	int		line;
 	int		column;
